refactor(core): refactor event system
