+++ packages/core/weapp/init/lifecycle.js
@@ -8,7 +8,6 @@ import { initRender } from './render';
 import { initData } from './data';
 import { initComputed } from './computed';
 import { initMethods } from './methods';
-import { initEvents } from './events';
 import { isArr, isFunc } from '../../shared/index';
 import Dirty from '../class/Dirty';
 import { WEAPP_APP_LIFECYCLE, WEAPP_PAGE_LIFECYCLE, WEAPP_COMPONENT_LIFECYCLE } from '../../shared/index';
@@ -153,8 +152,6 @@ export function patchLifecycle(output, options, rel, isComponent) {
 
       // created 不能调用 setData，如果有 dirty 在此更新
       vm.$forceUpdate();
-
-      initEvents(vm);
 
       Object.keys(outProps).forEach(k => (vm[k] = acceptProps[k]));
