+++ packages/core/weapp/init/methods.js
@@ -77,6 +77,10 @@ const proxyHandler = function(e) {
   }
   if (isFunc(fn)) {
     const paramsWithEvent = params.concat($event);
+    const customComData = e.detail._isCustomCom && e.detail.data;
+    if (customComData) {
+      paramsWithEvent.push(customComData);
+    }
     const hookRes = callUserHook(vm, 'before-event', {
       event: $event,
       params: paramsWithEvent
@@ -86,7 +90,7 @@ const proxyHandler = function(e) {
       // Event cancelled.
       return;
     }
-    return fn.apply(vm, params.concat($event));
+    return fn.apply(vm, paramsWithEvent);
   } else if (!model) {
     throw new Error('Unrecognized event');
   }
